requirements.txt
absl-py==1.0.0
asgiref==3.3.4
astunparse==1.6.3
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
joblib==1.0.1
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
mysqlclient==2.0.3
nltk==3.6.2
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.1
regex==2021.4.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.61.0
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
whitenoise==5.2.0
wrapt==1.13.3
zipp==3.6.0
